feat: move org allowlists to an Admin page, keep Integrations personal
In Teams mode the global Integrations page morphed into an org app-allowlist
policy editor and hid from plain members, which read as a different product per
role. The page is now the personal catalog for everyone in every mode, and the
org policy surfaces move to the Organization page, relabeled Admin:

- Admin (was Organization) gains Allowed integrations and Allowed AI models
  tabs (owner edits, admin read-only), shown only on a Teams host; the AI hub
  Workspace policy tab and the Integrations policy identity are removed.
- The Integrations sidebar entry is unconditional; canSeeIntegrationsPage,
  integrationsPageMode, and connectAffordance are deleted along with their
  two-destination routing in agent tabs, routine triggers, and Settings.
- Label rename only: ORGANIZATION_VIEW_ID and the organization/ dir stay.
- en/es/pt locales updated; wire surface GET/PUT /org/settings unchanged, the
  gateway remains the sole enforcer.
