+++ app/src/components/shell/sidebar-chrome.tsx
@@ -25,13 +25,11 @@ type ShellT = TFunction<["shell", "common", "portable", "teams"]>;
  *  optional Organization, Settings). */
 export function buildSidebarNavItems(args: {
   t: ShellT;
-  showIntegrations: boolean;
   showAiModels: boolean;
   showOrganization: boolean;
   setViewMode: (view: string) => void;
 }): SidebarNavItemEntry[] {
-  const { t, showIntegrations, showAiModels, showOrganization, setViewMode } =
-    args;
+  const { t, showAiModels, showOrganization, setViewMode } = args;
   return [
     {
       id: "dashboard",
@@ -40,17 +38,13 @@ export function buildSidebarNavItems(args: {
       onClick: () => setViewMode("dashboard"),
       dataAttrs: { "data-tour-target": "nav-dashboard" },
     },
-    ...(showIntegrations
-      ? [
-          {
-            id: INTEGRATIONS_VIEW_ID,
-            label: t("shell:sidebar.integrations"),
-            icon: <Blocks className="h-4 w-4" />,
-            onClick: () => setViewMode(INTEGRATIONS_VIEW_ID),
-            dataAttrs: { "data-tour-target": "nav-integrations" },
-          },
-        ]
-      : []),
+    {
+      id: INTEGRATIONS_VIEW_ID,
+      label: t("shell:sidebar.integrations"),
+      icon: <Blocks className="h-4 w-4" />,
+      onClick: () => setViewMode(INTEGRATIONS_VIEW_ID),
+      dataAttrs: { "data-tour-target": "nav-integrations" },
+    },
     ...(showAiModels
       ? [
           {
