+++ app/src/components/shell/sidebar.tsx
@@ -12,10 +12,7 @@ import { DEFAULT_TAB_ID } from "../../agents/standard-tabs";
 import { useCanCreateAgents } from "../../hooks/use-can-create-agents";
 import { useCapabilities } from "../../hooks/use-capabilities";
 import { useSidebarLayout } from "../../hooks/use-sidebar-layout";
-import {
-  canSeeAiModelsPage,
-  canSeeIntegrationsPage,
-} from "../../lib/org-roles";
+import { canSeeAiModelsPage } from "../../lib/org-roles";
 import { resolveAutoCollapse } from "../../lib/sidebar-auto-collapse";
 import { isTopLevelView } from "../../lib/top-level-views";
 import { useAgentStore } from "../../stores/agents";
@@ -65,10 +62,6 @@ export function Sidebar({ children }: { children: ReactNode }) {
   // only. Hidden entirely for plain members and single-player (canSeeOrganization
   // is the same gate as the members roster).
   const showOrganization = canSeeOrganization(capabilities);
-  // Teams v2: in a Teams workspace the Integrations page becomes org policy, so
-  // plain members lose both the page and its nav entry (they manage apps from
-  // the agent tab + Settings). Everyone else keeps it.
-  const showIntegrations = canSeeIntegrationsPage(capabilities);
   // Teams v2: in a Teams workspace the AI Models hub is owner/admin territory
   // (org-level provider credentials + admin model policy), so plain members lose
   // its nav entry too — they pick their model per agent in the composer.
@@ -189,7 +182,6 @@ export function Sidebar({ children }: { children: ReactNode }) {
           }
           navItems={buildSidebarNavItems({
             t,
-            showIntegrations,
             showAiModels,
             showOrganization,
             setViewMode,
