+++ app/src/components/tabs/use-routine-triggers.tsx
@@ -9,7 +9,6 @@ import { useTranslation } from "react-i18next";
 import { useIntegrationToolkits } from "../../hooks/queries/use-integrations";
 import { useAgentTriggerStatus } from "../../hooks/queries/use-triggers";
 import { useCapabilities } from "../../hooks/use-capabilities";
-import { canSeeIntegrationsPage } from "../../lib/org-roles";
 import type { Agent } from "../../lib/types";
 import { useUIStore } from "../../stores/ui";
 import { INTEGRATION_PROVIDER } from "../integrations/model";
@@ -43,16 +42,11 @@ export function useRoutineTriggers(
   const catalog = useIntegrationToolkits(INTEGRATION_PROVIDER, triggersEnabled);
 
   const setViewMode = useUIStore((s) => s.setViewMode);
-  const setSettingsSection = useUIStore((s) => s.setSettingsSection);
   const onReconnectTrigger = useCallback(() => {
-    // Same routing the Integrations tab's "Manage all" uses: the policy page
-    // when the caller can see it, else Settings > Connected accounts.
-    if (canSeeIntegrationsPage(capabilities)) setViewMode(INTEGRATIONS_VIEW_ID);
-    else {
-      setSettingsSection("connectedAccounts");
-      setViewMode("settings");
-    }
-  }, [capabilities, setViewMode, setSettingsSection]);
+    // Same routing the Integrations tab's "Manage all" uses: the global
+    // Integrations page, which everyone can reach.
+    setViewMode(INTEGRATIONS_VIEW_ID);
+  }, [setViewMode]);
 
   const triggerStatuses = useMemo(
     () => toStatusMap(statusQuery.data),
