+++ app/src/locales/pt/integrations.json
@@ -1,10 +1,5 @@
 {
   "title": "Integrações",
-  "policyPage": {
-    "subtitle": "Escolha quais aplicativos os agentes deste espaço de trabalho podem usar.",
-    "manageAccounts": "Gerencie seus aplicativos conectados",
-    "perAgentNote": "Quer limites mais restritos para um agente? Defina isso nas configurações desse agente."
-  },
   "loading": {
     "title": "Carregando suas integrações",
     "body": "Verificando seu provedor de integrações…"
