+++ app/src/locales/pt/settings.json
@@ -145,8 +145,7 @@
     "subtitle": "Aplicativos que você conectou. Escolha quais agentes podem usar cada um.",
     "heading": "Seus aplicativos",
     "empty": "Você ainda não conectou nenhum aplicativo.",
-    "connectMore": "Conectar mais aplicativos",
-    "connectHint": "Conecte novos aplicativos pela aba Integrações de qualquer agente."
+    "connectMore": "Conectar mais aplicativos"
   },
   "apiKeys": {
     "title": "Chaves de API",
